DCSW: Validate both CA variables are present and from the same CA (#35012)
<!-- Add the related story/sub-task/bug number, like Resolves #123, or
remove if NA -->

**Related issue:** Resolves #34973 

# Checklist for submitter

If some of the following don't apply, delete the relevant line.
## Testing

- [x] Added/updated automated tests
- [x] QA'd all new/changed functionality manually
