+++ server/service/mdm_test.go
@@ -1192,6 +1192,9 @@ func TestMDMWindowsConfigProfileAuthz(t *testing.T) {
 	ds.ValidateEmbeddedSecretsFunc = func(ctx context.Context, documents []string) error {
 		return nil
 	}
+	ds.GetGroupedCertificateAuthoritiesFunc = func(ctx context.Context, includeSecrets bool) (*fleet.GroupedCertificateAuthorities, error) {
+		return &fleet.GroupedCertificateAuthorities{}, nil
+	}
 
 	checkShouldFail := func(t *testing.T, err error, shouldFail bool) {
 		if !shouldFail {
@@ -1273,6 +1276,9 @@ func TestUploadWindowsMDMConfigProfileValidations(t *testing.T) {
 	}
 	ds.ValidateEmbeddedSecretsFunc = func(ctx context.Context, documents []string) error {
 		return nil
+	}
+	ds.GetGroupedCertificateAuthoritiesFunc = func(ctx context.Context, includeSecrets bool) (*fleet.GroupedCertificateAuthorities, error) {
+		return &fleet.GroupedCertificateAuthorities{}, nil
 	}
 
 	cases := []struct {
@@ -2387,155 +2393,6 @@ func TestBatchSetMDMProfilesLabels(t *testing.T) {
 		LabelsExcludeAny: []string{"baddy"},
 	}}, true, false, ptr.Bool(true), false)
 	require.NoError(t, err)
-}
-
-func TestValidateWindowsProfileFleetVariablesLicense(t *testing.T) {
-	t.Parallel()
-	profileWithVars := `<Replace>
-			<Item>
-				<Target>
-					<LocURI>./Device/Vendor/MSFT/Accounts/DomainName</LocURI>
-				</Target>
-				<Data>Host UUID: $FLEET_VAR_HOST_UUID</Data>
-			</Item>
-		</Replace>`
-
-	// Test with free license
-	freeLic := &fleet.LicenseInfo{Tier: fleet.TierFree}
-	_, err := validateWindowsProfileFleetVariables(profileWithVars, freeLic)
-	require.ErrorIs(t, err, fleet.ErrMissingLicense)
-
-	// Test with premium license
-	premiumLic := &fleet.LicenseInfo{Tier: fleet.TierPremium}
-	vars, err := validateWindowsProfileFleetVariables(profileWithVars, premiumLic)
-	require.NoError(t, err)
-	require.Contains(t, vars, "HOST_UUID")
-
-	// Test profile without variables (should work with free license)
-	profileNoVars := `<Replace>
-			<Item>
-				<Target>
-					<LocURI>./Device/Vendor/MSFT/Accounts/DomainName</LocURI>
-				</Target>
-				<Data>Static Value</Data>
-			</Item>
-		</Replace>`
-	vars, err = validateWindowsProfileFleetVariables(profileNoVars, freeLic)
-	require.NoError(t, err)
-	require.Nil(t, vars)
-}
-
-func TestValidateWindowsProfileFleetVariables(t *testing.T) {
-	tests := []struct {
-		name        string
-		profileXML  string
-		wantErr     bool
-		errContains string
-	}{
-		{
-			name: "no variables",
-			profileXML: `<Replace>
-				<Item>
-					<Target>
-						<LocURI>./Device/Vendor/MSFT/Policy/Config/System/AllowLocation</LocURI>
-					</Target>
-					<Data>1</Data>
-				</Item>
-			</Replace>`,
-			wantErr: false,
-		},
-		{
-			name: "HOST_UUID variable",
-			profileXML: `<Replace>
-				<Item>
-					<Target>
-						<LocURI>./Device/Vendor/MSFT/Policy/Config/System/AllowLocation</LocURI>
-					</Target>
-					<Data>$FLEET_VAR_HOST_UUID</Data>
-				</Item>
-			</Replace>`,
-			wantErr: false,
-		},
-		{
-			name: "HOST_UUID variable with braces",
-			profileXML: `<Replace>
-				<Item>
-					<Target>
-						<LocURI>./Device/Vendor/MSFT/Policy/Config/System/AllowLocation</LocURI>
-					</Target>
-					<Data>${FLEET_VAR_HOST_UUID}</Data>
-				</Item>
-			</Replace>`,
-			wantErr: false,
-		},
-		{
-			name: "multiple HOST_UUID variables",
-			profileXML: `<Replace>
-				<Item>
-					<Target>
-						<LocURI>./Device/Vendor/MSFT/Policy/Config/System/AllowLocation</LocURI>
-					</Target>
-					<Data>$FLEET_VAR_HOST_UUID-${FLEET_VAR_HOST_UUID}</Data>
-				</Item>
-			</Replace>`,
-			wantErr: false,
-		},
-		{
-			name: "unsupported variable",
-			profileXML: `<Replace>
-				<Item>
-					<Target>
-						<LocURI>./Device/Vendor/MSFT/Policy/Config/System/AllowLocation</LocURI>
-					</Target>
-					<Data>$FLEET_VAR_HOST_HARDWARE_SERIAL</Data>
-				</Item>
-			</Replace>`,
-			wantErr:     true,
-			errContains: "Fleet variable $FLEET_VAR_HOST_HARDWARE_SERIAL is not supported in Windows profiles",
-		},
-		{
-			name: "HOST_UUID with another unsupported variable",
-			profileXML: `<Replace>
-				<Item>
-					<Target>
-						<LocURI>./Device/Vendor/MSFT/Policy/Config/System/AllowLocation</LocURI>
-					</Target>
-					<Data>$FLEET_VAR_HOST_UUID-$FLEET_VAR_BOGUS_VAR</Data>
-				</Item>
-			</Replace>`,
-			wantErr:     true,
-			errContains: "Fleet variable $FLEET_VAR_BOGUS_VAR is not supported in Windows profiles",
-		},
-		{
-			name: "unknown Fleet variable",
-			profileXML: `<Replace>
-				<Item>
-					<Target>
-						<LocURI>./Device/Vendor/MSFT/Policy/Config/System/AllowLocation</LocURI>
-					</Target>
-					<Data>${FLEET_VAR_UNKNOWN_VAR}</Data>
-				</Item>
-			</Replace>`,
-			wantErr:     true,
-			errContains: "Fleet variable $FLEET_VAR_UNKNOWN_VAR is not supported in Windows profiles",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// Pass a premium license for testing (we're not testing license validation here)
-			premiumLic := &fleet.LicenseInfo{Tier: fleet.TierPremium}
-			_, err := validateWindowsProfileFleetVariables(tt.profileXML, premiumLic)
-			if tt.wantErr {
-				require.Error(t, err)
-				if tt.errContains != "" {
-					require.Contains(t, err.Error(), tt.errContains)
-				}
-			} else {
-				require.NoError(t, err)
-			}
-		})
-	}
 }
 
 func androidConfigProfileForTest(t *testing.T, name string, content map[string]any, labels ...*fleet.Label) *fleet.MDMAndroidConfigProfile {
